fix: clean sweep wandb integration and misc utilities
