+++ metta/sweep/protein_wandb.py
@@ -1,13 +1,15 @@
+import json
 import logging
 import time
 from copy import deepcopy
 from datetime import datetime, timezone
-from typing import Any, Tuple
+from typing import Any, Tuple, cast
 
 import pufferlib
 import wandb
 
 from metta.common.util.numpy_helpers import clean_numpy_types
+from metta.common.util.retry import retry_on_exception
 
 from .protein import Protein
 
@@ -33,8 +35,9 @@ class WandbProtein:
             wandb_run (wandb.Run, optional): The wandb run to use. If None, uses the current run.
             max_runs_to_load (int, optional): Maximum number of previous runs to load. Defaults to 100.
         """
-        self._wandb_run = wandb_run or wandb.run
+        self._wandb_run = wandb_run or None
         assert self._wandb_run is not None, "No active wandb run found"
+        self._wandb_run = cast(Any, self._wandb_run)
 
         # Get sweep ID - handle both sweep_id attribute and sweep.id fallback
         self._sweep_id = getattr(self._wandb_run, "sweep_id", None)
@@ -68,7 +71,7 @@ class WandbProtein:
         # Generate protein suggestion - let exceptions propagate
         self._generate_protein_suggestion()
 
-        # CRITICAL: Overwrite WandB agent's suggested parameters with Protein's suggestions
+        # Overwrite WandB agent's suggested parameters with Protein's suggestions
         wandb_config = self._transform_suggestion(deepcopy(self._suggestion))
         # Unlock config to allow overwriting WandB agent's suggestions
         self._wandb_run.config.__dict__["_locked"] = {}
@@ -96,7 +99,7 @@ class WandbProtein:
             }
         )
 
-        # CRITICAL: Also record the observation with the Protein for future runs
+        # Also record the observation with the Protein for future runs
         logger.info(f"Recording observation: {self._suggestion}, {objective}, {cost}")
         self._protein.observe(self._suggestion, objective, cost, False)
 
@@ -115,7 +118,7 @@ class WandbProtein:
             }
         )
 
-        # CRITICAL: Also record the failure with the Protein
+        # Also record the failure with the Protein
         self._protein.observe(self._suggestion, 0.0, 0.0, True)
 
     def suggest(self, fill=None) -> Tuple[dict[str, Any], dict[str, Any]]:
@@ -132,18 +135,17 @@ class WandbProtein:
         try:
             # First try with just numpy cleaning
             cleaned = clean_numpy_types(suggestion)
-            # Test if it's JSON serializable
-            import json
 
+            # Test if it's JSON serializable
             json.dumps(cleaned)
             return cleaned
+
         except (TypeError, ValueError):
             # If that fails, do a deep conversion to basic Python types
-            return self._deep_clean(suggestion)
+            return cast(dict[str, Any], self._deep_clean(suggestion))
 
     def _deep_clean(self, obj):
         """Recursively convert any object to JSON-serializable Python types."""
-        import json
 
         if isinstance(obj, dict):
             # Already a regular dict, just recursively clean values
@@ -180,86 +182,69 @@ class WandbProtein:
         except Exception as e:
             logger.warning(f"Could not load previous runs: {e}")
 
+    @retry_on_exception(max_retries=3, retry_delay=2)
     def _get_runs_from_wandb(self) -> list:
         # Skip loading runs if this is not a real sweep or in offline mode
         if not self._sweep_id:
             return []
 
-        # Add retry logic for network calls
-        max_retries = 3
-        retry_delay = 2
+        start_time = time.time()
+        logger.info(f"Starting API call to fetch runs for sweep {self._sweep_id}")
+        logger.info(f"Project path: {self._wandb_run.entity}/{self._wandb_run.project}")
 
-        for attempt in range(max_retries):
-            try:
-                start_time = time.time()
-                logger.info(f"Starting API call to fetch runs for sweep {self._sweep_id}")
-                logger.info(f"Project path: {self._wandb_run.entity}/{self._wandb_run.project}")
+        runs = self._api.runs(
+            path=f"{self._wandb_run.entity}/{self._wandb_run.project}",
+            filters={
+                "sweep": self._sweep_id,
+                # Temporarily comment out to test if this is causing slowness
+                # "summary_metrics.protein.state": {"$exists": True},
+                # "id": {"$ne": self._wandb_run.id},
+            },
+            order="+created_at",
+        )
 
-                # Try a simpler filter first to see if that's the issue
-                # Just filter by sweep ID initially
-                runs = self._api.runs(
-                    path=f"{self._wandb_run.entity}/{self._wandb_run.project}",
-                    filters={
-                        "sweep": self._sweep_id,
-                        # Temporarily comment out to test if this is causing slowness
-                        # "summary_metrics.protein.state": {"$exists": True},
-                        # "id": {"$ne": self._wandb_run.id},
-                    },
-                    order="+created_at",
-                )
+        api_call_time = time.time() - start_time
+        logger.info(f"API call completed in {api_call_time:.2f} seconds")
 
-                api_call_time = time.time() - start_time
-                logger.info(f"API call completed in {api_call_time:.2f} seconds")
+        # OPTIMIZATION: Process runs lazily and limit the number fetched
+        # This avoids fetching thousands of runs at once
+        result_runs = []
+        max_runs = self._max_runs_to_load
 
-                # OPTIMIZATION: Process runs lazily and limit the number fetched
-                # This avoids fetching thousands of runs at once
-                result_runs = []
-                max_runs = self._max_runs_to_load
+        logger.info(f"Fetching up to {max_runs} previous runs from sweep {self._sweep_id}")
 
-                logger.info(f"Fetching up to {max_runs} previous runs from sweep {self._sweep_id}")
+        fetch_start = time.time()
+        for _, run in enumerate(runs):
+            # Apply filters manually after fetching
+            if run.id == self._wandb_run.id:
+                continue
 
-                fetch_start = time.time()
-                for _i, run in enumerate(runs):
-                    # Apply filters manually after fetching
-                    if run.id == self._wandb_run.id:
-                        logger.debug(f"Skipping current run {run.id}")
-                        continue
+            if not run.summary.get("protein.state"):
+                continue
 
-                    if not run.summary.get("protein.state"):
-                        logger.debug(f"Skipping run {run.id} - no protein state")
-                        continue
+            if len(result_runs) >= max_runs:
+                logger.info(f"Reached maximum of {max_runs} runs, stopping fetch")
+                break
 
-                    if len(result_runs) >= max_runs:
-                        logger.info(f"Reached maximum of {max_runs} runs, stopping fetch")
-                        break
+            result_runs.append(run)
 
-                    result_runs.append(run)
+        fetch_time = time.time() - fetch_start
+        logger.info(f"Successfully fetched {len(result_runs)} runs in {fetch_time:.2f} seconds")
+        return result_runs
 
-                    # Log progress every 10 runs to show it's working
-                    if (len(result_runs)) % 10 == 0:
-                        logger.info(f"Fetched {len(result_runs)} runs...")
+    def _validate_run(self, run):
+        """
+        Validate a run. Note that a failed run is not necessarily invalid.
 
-                fetch_time = time.time() - fetch_start
-                logger.info(f"Successfully fetched {len(result_runs)} runs in {fetch_time:.2f} seconds")
-                logger.info(f"Total time: {time.time() - start_time:.2f} seconds")
-                return result_runs
+        Args:
+            run (wandb.Run): The run to validate.
 
-            except Exception as e:
-                logger.warning(f"API call failed (attempt {attempt + 1}/{max_retries}): {e}")
-                logger.warning(f"Exception type: {type(e).__name__}")
-                if attempt < max_retries - 1:
-                    time.sleep(retry_delay)
-                    continue
-                else:
-                    logger.error(f"Failed to fetch runs after {max_retries} attempts")
-                    return []
-
-    def _update_protein_from_run(self, run):
-        logger.info(f"Processing run {run.name} (ID: {run.id})")
-
+        Returns:
+            bool: True if the run is valid, False otherwise.
+        """
         if run.summary.get("protein.state") == "initializing":
             logger.info(f"Skipping run {run.name} - still initializing")
-            return
+            return False
 
         if run.summary.get("protein.state") == "running":
             # Handle both formats: with and without microseconds
@@ -269,30 +254,40 @@ class WandbProtein:
             except ValueError:
                 # Try without microseconds
                 last_hb = datetime.strptime(heartbeat_str, "%Y-%m-%dT%H:%M:%SZ").replace(tzinfo=timezone.utc)
+
             if (datetime.now(timezone.utc) - last_hb).total_seconds() > 5 * 60:
                 self._defunct += 1
                 logger.info(f"Skipping run {run.name} - defunct (no heartbeat for >5 min)")
-                return
-            self._num_running += 1
-            logger.info(f"Skipping run {run.name} - still running")
-            return
 
-        try:
-            suggestion, info = self._suggestion_from_run(run)
-            logger.info(f"Extracted suggestion from run {run.name}: {suggestion}")
-        except Exception as e:
-            logger.warning(f"Failed to extract suggestion from run {run.name}: {e}")
-            self._invalid += 1
-            return
+            else:
+                self._num_running += 1
+                logger.info(f"Skipping run {run.name} - still running")
 
-        # Check for duplicate "running" state check
-        if run.summary.get("protein.state") == "running":
-            self._num_running += 1
-            return
+            return False
 
+        return True
+
+    def _get_suggestion_data_from_run(self, run):
+        """
+        Get the suggestion data from a run.
+
+        Args:
+            run (wandb.Run): The run to get the suggestion data from.
+
+        Returns:
+            tuple: A tuple containing the suggestion, info, objective, cost, and is_failure.
+        """
+        suggestion, info = self._suggestion_from_run(run)
         objective = run.summary.get("protein.objective", 0)
         cost = run.summary.get("protein.cost", 0)
         is_failure = run.summary.get("protein.state") == "failure"
+        return suggestion, info, objective, cost, is_failure
+
+    def _update_protein_from_run(self, run):
+        if not self._validate_run(run):
+            return
+
+        suggestion, info, objective, cost, is_failure = self._get_suggestion_data_from_run(run)
 
         if is_failure:
             self._num_failures += 1
@@ -314,86 +309,23 @@ class WandbProtein:
         # The suggestion from WandB is in nested format: {"trainer": {"optimizer": {"learning_rate": 0.0005}}}
         # But protein.observe() expects flattened format: {"trainer/optimizer/learning_rate": 0.0005}
         try:
-            # Skip empty suggestions
-            if not suggestion:
-                logger.warning(f"Empty suggestion from run {run.name}, skipping observation")
-                self._invalid += 1
-                return
+            # Convert to dict if it's not already
+            suggestion_dict = dict(suggestion)
 
-            # Check suggestion type before flattening
-            if not isinstance(suggestion, dict):
-                # Handle WandB's SummarySubDict and other dict-like objects
-                try:
-                    suggestion = dict(suggestion)
-                except Exception as e:
-                    logger.warning(f"Failed to convert suggestion from run {run.name} to dict: {e}")
-                    logger.debug(f"Suggestion type: {type(suggestion)}")
-                    self._invalid += 1
-                    return
-
-            logger.info(f"Original suggestion structure: {suggestion}")
-            logger.info(f"Suggestion type: {type(suggestion)}, is dict: {isinstance(suggestion, dict)}")
-
-            # Ensure we have a proper dict before flattening
-            if hasattr(suggestion, "items"):
-                # Convert dict-like objects to regular dict
-                suggestion_dict = dict(suggestion)
-                logger.info(f"Converted to dict: {suggestion_dict}")
-            else:
-                suggestion_dict = suggestion
-
+            # Flatten the suggestion
             flattened_suggestion = dict(pufferlib.unroll_nested_dict(suggestion_dict))
-            logger.info(f"Flattened suggestion: {flattened_suggestion}")
-
-            # Validate flattened suggestion is not empty
-            if not flattened_suggestion:
-                logger.warning(f"Flattened suggestion from run {run.name} is empty. Original: {suggestion}")
-                self._invalid += 1
-                return
 
             # Pass flattened suggestion to Protein
             self._protein.observe(flattened_suggestion, objective, cost, is_failure)
-            logger.info(f"Successfully recorded observation from run {run.name}: objective={objective}, cost={cost}")
 
-        except AssertionError as e:
-            # Handle missing parameters gracefully
-            if "Missing hyperparameter" in str(e):
-                logger.warning(f"Run {run.name} has incomplete parameters: {e}")
-                logger.debug(f"Available parameters: {list(flattened_suggestion.keys())}")
-                try:
-                    logger.debug(f"Expected parameters: {list(self._protein.hyperparameters.flat_spaces.keys())}")
-                except (AttributeError, TypeError):
-                    logger.debug("Expected parameters: <not available>")
-                # Skip this observation as it's incompatible with current config
-                self._invalid += 1
-            else:
-                # Re-raise other assertion errors
-                raise
-        except KeyError as e:
-            # Handle missing keys in hyperparameter spaces
-            logger.warning(f"Run {run.name} has parameter not in current config: {e}")
-            logger.debug(f"Flattened suggestion keys: {list(flattened_suggestion.keys())}")
-            try:
-                logger.debug(f"Expected parameters: {list(self._protein.hyperparameters.flat_spaces.keys())}")
-            except (AttributeError, TypeError):
-                logger.debug("Expected parameters: <not available>")
-            self._invalid += 1
         except Exception as e:
-            logger.warning(f"Failed to record observation in Protein for run {run.name}: {type(e).__name__}: {e}")
-            logger.debug(f"Suggestion that failed: {suggestion}")
-            logger.debug(f"Flattened suggestion: {flattened_suggestion}")
-            logger.debug(f"Exception type: {type(e).__name__}")
-            logger.debug(f"Exception details: {str(e)}")
-            # Try to provide more context about the failure
-            try:
-                logger.debug(f"Objective: {objective}, Cost: {cost}, Is_failure: {is_failure}")
-            except Exception:
-                pass
-            # Still increment invalid count since we couldn't learn from this observation
+            logger.warning(f"Failed to record observation from run {run.name}: {type(e).__name__}: {e}")
             self._invalid += 1
+            return
 
         # Clean the info to remove any WandB objects before storing
-        self._suggestion_info = self._deep_clean(info)
+        cleaned_info = self._deep_clean(info)
+        self._suggestion_info = cast(dict[str, Any], cleaned_info)
 
     def _suggestion_from_run(self, run):
         """
@@ -402,20 +334,8 @@ class WandbProtein:
         # Try to get the flattened suggestion items first (newest format)
         suggestion_items = run.summary.get("protein.suggestion_flattened_items", None)
 
-        if suggestion_items is not None and isinstance(suggestion_items, list):
-            # Reconstruct the flattened dict from the list of tuples
-            suggestion = dict(suggestion_items)
-            logger.info(f"Reconstructed flattened suggestion from items for run {run.name}")
-        else:
-            # Try the old flattened format (which WandB auto-nests)
-            suggestion = run.summary.get("protein.suggestion_flattened", None)
-
-            if suggestion is not None:
-                logger.info(f"Found flattened suggestion for run {run.name} (may be auto-nested by WandB)")
-            else:
-                # Fallback to nested format (oldest format)
-                suggestion = run.summary.get("protein.suggestion", {})
-                logger.info(f"Using nested suggestion format for run {run.name} (legacy)")
+        # Reconstruct the flattened dict from the list of tuples
+        suggestion = dict(suggestion_items)
 
         # Handle case where WandB stored the suggestion as a string
         if isinstance(suggestion, str):
@@ -473,6 +393,7 @@ class WandbProtein:
         # But we'll also store the flattened version explicitly
         cleaned_suggestion = self._deep_clean(self._suggestion)
         cleaned_info = self._deep_clean(info)
+        self._suggestion_info = cast(dict[str, Any], cleaned_info)
 
         # Save both nested and flattened versions to wandb summary
         # This ensures backward compatibility while fixing the loading issue
@@ -487,31 +408,3 @@ class WandbProtein:
                 "protein.suggestion_info": cleaned_info,
             }
         )
-
-
-def create_wandb_sweep(sweep_name: str, wandb_entity: str, wandb_project: str):
-    """
-    Create a new wandb sweep with a dummy parameter (Protein will control all suggestions).
-
-    Args:
-        sweep_name (str): The name of the sweep.
-        wandb_entity (str): The wandb entity (username or team name).
-        wandb_project (str): The wandb project name.
-
-    Returns:
-        str: The ID of the created sweep.
-    """
-    sweep_id = wandb.sweep(
-        sweep={
-            "name": sweep_name,
-            "method": "bayes",  # This won't actually be used since we override suggestions
-            "metric": {"name": "protein.objective", "goal": "maximize"},
-            "parameters": {
-                # WandB requires at least one parameter, but Protein will override all suggestions
-                "dummy_param": {"values": [1]}
-            },
-        },
-        project=wandb_project,
-        entity=wandb_entity,
-    )
-    return sweep_id
